+++ src/main/java/ku_rum/backend/global/exception/user/DuplicateEmailException.java
@@ -1,6 +1,6 @@
 package ku_rum.backend.global.exception.user;
 
-import ku_rum.backend.global.response.status.ResponseStatus;
+import ku_rum.backend.global.support.response.status.ResponseStatus;
 import lombok.Getter;
 
 @Getter
